adds `generic_typedefs` parameter to yaml file and also fixes typedef resolution in generic parameters. Issues: #1 Non expanded typedefs in generics ``` typedef NSString * NSReadingOptionKey NSDictionary<NSReadingOptionKey, id> ``` was translated into `NSDictionary<NSReadingOptionKey, ?>` and was not able to compile NSReadingOptionKey is unknown. ```
#2 Special case type resolution for generics
Once issue #1 was fixed it was resolved too deeply into `NSDictionary<String, ?>`. And this doesn't work as String is not subclass of NSObject.
And it gets to String as resolution path is following: `NSReadingOptionKey-> NSString * -> String`
To solve this special configuration key is introduced to resolve types specially for generics: `generic_typedefs` and now NSString* typedef looks as below in foundation.yaml:
```
typedefs:
    'NSString *': String
generic_typedefs:
    'NSString *': NSString
```

Which will produce String for general usage and to NSString for generic parameter resolution.